--- a/src/chat-heads.js
+++ b/src/chat-heads.js
@@ -1,23 +1,25 @@
-import React from 'react'
+import React, { PureComponent } from 'react'
 
-const ChatHeads = props => {
-  return (
-    <ul className="chat-heads">
-      {props.users.map(user => {
-        return (
-          <li key={user.id}>
-            <div
-              className="chat-head-img"
-              style={{
-                backgroundImage: `url(${user.avatar})`
-              }}
-            />
-            <div className="chat-head-name">{user.name}</div>
-          </li>
-        );
-      })}
-    </ul>
-  )
+class ChatHeads extends PureComponent {
+  render() {
+    return (
+      <ul className="chat-heads">
+        {this.props.users.map(user => {
+          return (
+            <li key={user.id}>
+              <div
+                className="chat-head-img"
+                style={{
+                  backgroundImage: `url(${user.avatar})`
+                }}
+              />
+              <div className="chat-head-name">{user.name}</div>
+            </li>
+          );
+        })}
+      </ul>
+    );
+  }
 }
 
 export default ChatHeads;