## Convert components to instances of `React.PureComponent`

React.PureComponent implements `shouldComponentUpdate` behind the scenes so we should see some performance improvements after the initial mount and render.

Here are the new stats

**Mount**
Functional: 7.98 ms
PureComponent: 8.33 ms (new)

**Initial render before**
Functional: 18.8 ms
PureComponent: 20.1 ms (new) 

**First update**
Functional: 8 ms update
PureComponent: 1.7 ms (new)

**Update while typing in input**
Functional: 3.5 ms
PureComponent: 1.5 ms (new)